Fix vendoring
There is a rule for go build: all sources files of a given package must
be in the directory of that package. Not in subdirectories.
Which means that {#include "hidapi/mac/hid.c"} is not correct as hid.c
is not in the package directory, so "go build" will not notice its
changes and go mod -vendor will not work (https://github.com/golang/go/issues/26366).

So this patch creates 3 files (hidapi_linux.h, hidapi_windows.h,
hidapi_mac.h) which are go-generated by internal/gen.go. The generator
just inlines any includes from the libusb/ and hidapi/ directories.
This means that the libusb/ and hidapi/ are not needed anymore for
distribution (and for go mod -vendor): they are used only for the "go
generate" phase.